refactor(i18n): 将本地化代码集中为 src/i18n 模块
