--- src/i18n.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated 请使用 './common/i18n' 代替。
- * 此文件保留以便平滑迁移，将在后续版本中移除。
- */
-export * from './common/i18n';
