+++ src/i18n/core.ts
