+++ src/test/modules/userFacingErrors.test.ts
@@ -4,7 +4,7 @@
 
 import * as assert from 'assert';
 import { getUserFacingErrorMessage } from '../../modules/userFacingErrors';
-import { __setLanguageOverrideForTests } from '../../common/i18n';
+import { __setLanguageOverrideForTests } from '../../i18n';
 
 suite('Modules — UserFacingErrors', () => {
